refactor(PDFProView): :truck: data & utils moved to root src folder
