+++ src/data/ocr-sample.ts
